Merged PR 503217: disable manual check for ping-pong timeout
The TCP keep alive is supported in the current libwebsocket version 2.4.1, thus no need to run manual check.
The manual check is not robust anyway. When sending a large amount of data in one-shot to the socket connection, the choked state might keep longer time than the PING_PONG_INTERVAL_SECONDS (but this is not due to network congestion), and thus the connection is closed unexpectedly.

Related work items: #1087619