Merged PR 515408: Send End-of-Audio to service; add timestamp in log
CRIS expects the end-of-audio signal to detect completion of audio data, otherwise it takes long time to detect speech.end (5s in interactive mode, and until connection timeout in conversation and dictation mode). Send end-of-audio to service when reaching end of file. It is open how to detect that when using microphone. Bing speech works fine (takes a little time, ca. 2s to detect end).

Related work items: #1089052, #1098917, #1103451