Merged PR 516645: Fix a bug in AudioStreamFlush(); Add __FUNCTION__ to LogInfo
AudioStreamFlush() checks whether the Websocket connection is open, before sending the zero-sized buffer. It is not needed, since here it only puts the request into sending io queue. And it leades to bug, if the audio to be sent is very short, and then the succeeding AudioStreamFlush() (to send end-of-audio message to service) will fail.

Related work items: #1105406