Merged PR 530275: Add user-defined messages and CDSDK endpoint to USPLib
User-defined messages consist of a user defined path and a payload. I envision that if the message cannot be sent immediately, it should be the job of the USP client to queue messages, not the UspLib.

Also added support for the CDSDK endpoint. This involved adding a new UspAuthenticationType, and adding back the user agent.

Related work items: #1096481, #1127503