Merged PR 537046: Add USP support for receiving an audio stream
Added UspOnAudioStreamStart callback, which is called when the first chunk of an audio stream is received from the service. Subsequent chunks are then added to an IoBuffer that is passed along on the initial callback. Added TTS option to the uspclientconsole.

Also added a request header: "X-Output-AudioCodec" which tells the service to send the audio down as a WAV.

Related work items: #1135210