Merged PR 595251: Throw proper exceptions
As discussed, throwing proper exceptions (recording the callstack the the throw-site) inside the carbon dll instead of opaque ints. At the border, catching them, storing in the handle table and returning corresponding handles as the return value. In the cpp layer, using the provided c-api functions to retrieve error details and re-throw as std::runtime_error.

Related work items: #1216258