Merged PR 621711: Fixing a deadlock between usp and usp adapter using object with site
Fixing a deadlock between usp and usp adapter using object with site. Also some of the locks in the adapter were made shorter, so that we never call the callback under a lock. I also removed explicit disconnect from C#, because the recognizer should already do that in the correct order.

Related work items: #1243368