Merged PR 623737: Making usp adapter callback safe if shutdown is in parallel with callback
Currently adapter is using GetSite without checking whether the site actually exists.
This leads to crashes in case the delete has already been called on site, but the USP thread is still in process of propagating a message up the stack.
In order to avoid that we have to make sure that all callbacks are implemented in a pattern specified in ISpxObjectWithSiteInitImpl:
 - firstly lock the weak pointer on stack
 - if the pointer still valid, you are allowed to call the site
 - perform the call
 - the on stack pointer is destroyed

This guarantees that all objects in the hierarchy are still alive during callback execution.

Related work items: #1229396