Merged PR 646457: Fire nomatch result when response from service is timed out; Disabled disconnect in python test as the old bug still exists.
In case of service does not return any response within the timeout value, the client just passes an empty result to users. And when users try to retrieve result, they get exception. The change fixed the bug and provides the following behanvior
 - in core, when wait_for_recognition() sees timeout, it calls EnsureFireResult() which fires the NoMatch result (this is actually not correct, but it is the way currently being implemented for the other client errors too).
 - if no result is received, it returns SPXERR_TIMEOUT instead of empty result, so that users get error at the point when the error occurs.

The bug 1180542 (Deadlock in Disconnect) has not been resolved yet, so disable the disconnect in python test for now.

Related work items: #1269084