Merged PR 687229: Remove USP DNS cache thread
Currently the DNS cache implementation within the USP is only being used for Linux, however the object is still being created on every platform.

During the lifetime of the object, a DNS worker thread is spawned.

Only creating the object on Linux now, to remove unnecessary thread on other platforms.