Merged PR 700765: Java: exposing Audio synthesis data and status
exposes the raw audio data and synthesis status to java clients.

note: one can argue if this should be a stream instead of an array. but for now, this is as is cpp

Note: due to a swig issue, i had to revert the naming changes for the internal classes from java, to "original" (i.e., capital first letter). This should not be an issue though since this is only visible in the internal classes, not the client facing apis

Related work items: #1329761