Merged PR 733744: Fixing reconnection logic
Due to reconnect, the state machine of the USP adapter was reset to idle and the new workflow started from scratch. This caused problems on the upper layer (audio session) because it is expecting messages in a particular order: each StartTurn should be complemented by EndTurn. Now depending on the current state of the USP adapter we make sure that Start/EndTurn is communicated properly.

Related work items: #1367526