Merged PR 749044: USP Client Console: output session ID
For correlation in case of crashes / hangs. To see those in VSTS, go to the test tab of a build.
Find the file that has the test you're interested in (e.g., `uspclientconsole - Linux-x64-Debug`) and download it's attachment (here: `test-uspconsole.xml`). In the XML, session IDs are logged for each test.

Related work items: #1386428