Merged PR 766847: Creating Xcode targets to build for iOS and macOS, with the iOS target being a universal framework for x86_64 and arm64.
Creating Xcode targets to build for iOS and macOS, with the iOS target being an universal framework for x86_64 and arm64.
Zipping the frameworks in the format they will be provided as CocoaPods.
Adding the iOS and OSX .podspecs for the CocoaPods that will be distributed for the Speech SDK.
Configuring the SpeechSDK_iOS ans SpeechSDK Test apps to use the .podspec that is currently configured to run locally.
Creating Xcode targets to build for iOS and macOS, with the iOS target being a universal framework for x86_64 and arm64.

Related work items: #1408492, #1408495