Pull Request 775889: [GA, rc2] On failure of a connection promise, raise cancel event instead of eating error
There was a broken chain of promises returned in Recognize (the internal method) that resulted in rejected promises not rolling up and out of the method. This allowed failures in methods on the far side of the break to fail and not report anything out.
This change connects the broken chain and then adds a new (internal) event generation for "internal error".
Each recognizer type then handles that error by raising the cancel event and calling it's continuation callback (if passed).
Also expanded existing tests to better report errors, but sadly don't have a way to force a failure at the stage where the failures would have been lost. Did test with a modified connection factory to validate the errors now get surfaced.