Merged PR 779349: [post-GA] check whether C++ result handle is valid to avoid access violation
If we still want to re-build bits for 1.0, this one should be merged.

If an exception is thrown during recognizeOnceAsync, the result handle from C++ could be null. The recent CancellationDetails API requires a valid C++ result. Add this check to avoid access violation in Objective C, in case an exception is thrown during recognizing and user want to access CancellationDetails.
This a workaround. The better solution to is return error/exceptions to users, instead of an canceled result without details.

Related work items: #1375701