Merged PR 784333: Bug: validate strings on various SpeechConfig::From*()
please note that we have a problem with the C++ API,

in particular, one can write

auto config = SpeechConfig::FromSubscription(nullptr, nullptr);

the problem with this is that there is an implicit conversion to std::string happening on the parameter, however, std::string(nullptr) is undefined behavior - and crashes the app!!

this will be a surprise for c++ programmers, especially since they could pass this through a variable, so the nullptr is not explicitly visible.

for now, I excluded nullptr-tests from the tests for the above reason.

Related work items: #1441056