Merged PR 780448: Adding a sample test project for testing the installation of the SpeechSDK as a CocoaPod
The logic of the app is the same as the 'helloworld' quickstart project. The xcodeproject is configured to use CocoaPods as specified in the Podfile, but 'pod install' needs to be run first in order for the pod to correctly be configured and added to the sample project. The Readme.txt file has more information regarding this.

Related work items: #1437418