Merged PR 791662: Handle audio thread exceptions properly
Fix the invalid endpoint problem in USP module. It calls the Error on audio session to send out a canceled event.

If there is any exception happens in the audio pump thread, call the site of audio pump, which is also the audio session, to shot out a canceled event.

In case of any exception in other threads in audio session, send out a canceled event.

The audio session state is not sent to Error state. This is after lengthy discussion with Rob. We decided not to. Reasons are it would take too long to add a Error state correctly. Also, we are refactoring state management of audio session.

After exceptions, if users call startRecogOnce or startContinuous, we will throw invalid_transit exception. The users can't use the recognizer anymore. This is more or less the same as marking Error state. The users have to create another recognizer.

Related work items: #1394625