Merged PR 804792: Support multiple concurrent requestIds in the USP
*** This PR has no functional effect on any existing Carbon scenarios. ***

This change enables partial-duplex support within the USP client layer. In this context partial-duplex means the ability to send messages up asynchronously.

The core of this change involves allowing the USP to keep track of multiple requestIds at once. A requestId is generated at the start of a new request. All messages for a given request are sent with the same requestId. The client only accepts messages received from the service if they contain a requestId matching one on the client. The client invalidates a requestId once a turn end message is received with a matching requestId.

There is one telemetry change worth noting:
The deviceId was moved from the first request's telemetry payload into the connection telemetry payload (which is sent right before the first request).

Here is a wiki page with more context:
https://cognitionwiki.com/display/CSACAI/USP+Requests+and+the+X-RequestId+header

Related work items: #1118147