Merged PR 828249: Throwing Exception When Recognizer is Disposed in Async Recognition
If any continuous recognition or RecognizeOnceAsync is not awaited in the main thread, the main thread can dispose a respective recognizer before async task has been completed.

Added a lock to tasks and base Dispose function, and checked whether the recognizer has already been disposed. If it is, currently running task throws ObjectDisposedException and stops instead of crashing the app.

Related work items: #1351305