Merged PR 856466: disable sending connection telemetry message
Currently, the connection telemetry message is sent to service without RequestId. This makes Bing Speech/Intent service close the connection. As workaround, since we do not use this telemetry data for now, we disable sending this telemetry data.
When we come back to revisit the telemetry, we need to have a proper design to send the connection data with requestId. See VSO item for details: https://msasg.visualstudio.com/Skyman/_workitems/edit/1530463.

Related work items: #1531286