Merged PR 852047: Speech Recognition Test in Interactive Mode
The speech recognition results in the interactive mode and while doing continuous recognition are not the same.  It is an expected behavior in the SDK side.

When a long file is sent for a recognition using RecognizeOnceAsync, it is divided into terms in the server side. Each term has 15 seconds of recognition limitation.

By checking the starting time offsets and durations of those terms and matching with the local file, apparently, there are cases that a term ends without the last word completely pronounced. In this case, some part of the word is tried to be predicted and returns either something does not exist in the test data or as an empty string.

For instance, recognition of batman.wav: "inexhaustible wealth"
the term ends: "inexhaust"
the new term starts from: "ible wealth"
recognition: "possible" "wealth"
in continuous recognition and in the test data: "inexhaustible wealth".

Related work items: #1409006