Merged PR 863562: Enable non-browser based websocket connections in JavaScript
While the unit test engine that the JavaScript code has been using runs as part of NodeJs, it also emulates a browser's script engine while doing so, exposing primitives that do not exist in NodeJs, but do in browser based script engines.

Like WebSockets.

This change adds support for NodeJs execution and uses and existing websocket implementation.

As the WebSocket implementation is no longer part of the test environment when run with the separate implementation, a host of timing bugs were exposed, both in test cases and the JavaScript implementation. Fixes are included.

The tests are also now configured to run with both the browser emulated WebSockets and the npm package variant via a test hook.

Related work items: #1544406