Merged PR 865074: Ensure errors druing audio transmission cancel reco.
Ensure errors druing audio transmission cancel reco.

There was a bug where when sending audio to the service fails because of bad data being sent in, the error was not properly propagating to the caller and it looked like a hang.

Now the error will propagate to the caller.

Also found that when an error forced a cancellation, only the cancel method was getting called, and not the continuation for RecognizeOnceAsync.

Related work items: #1545856