Merged PR 868030: Add per test case name logging to JS Tests
With the addition of multiple passes over the tests for each websocket implementation, it's gotten increasingly difficult to tell what stdout spew came from what test, and jest doesn't make it easy to find out.

So adding the name of each test, and a start / stop timer to each.

Related work items: #1551338