Merged PR 871380: [Release 1.2.0] Not throw exception on Disconnect() if recognizer is gone
Connection can be destroyed after the associated recognizer (e.g. in C#). In this case, when the connection is released, it calls DisconnectAll() on both connected and disconnected events, which cause exceptions since recognizer is already gone. Although the exception has been caught during destructor (so there is no direct impact on users), users might see that an exception has been thrown inside SDK if he uses VS debug and turn all exception notification. This could confuse users. The change avoids to throw such exceptions. It is ok to do so, since after the recognizer is gone, there is no point to call Disconnect on that recognizer.

Related work items: #1554141