Merged PR 895312: Leaking Debug logs on Release
SPX_CONFIG_INCLUDE_ALL is defined both in debug and release mode:

```
#ifdef _DEBUG
#define SPX_CONFIG_INCLUDE_ALL_DBG 1
#define SPX_CONFIG_INCLUDE_ALL 1
#else
#define SPX_CONFIG_INCLUDE_ALL 1
#endif
```

in spxdebug.h header, there are separate directives for debug and release logs and SPX_CONFIG_INCLUDE_ALL is used to differentiate those. But there are not separate directives for the cases that reports failures and exits with errors. Having condition of
 `#if defined(SPX_CONFIG_TRACE_EXITFN_ON_FAIL) || defined(SPX_CONFIG_INCLUDE_ALL) || defined(SPX_CONFIG_INCLUDE_ALL_DBG)`
causes the failure logs to be printed on release mode as well.

Related work items: #1556846