Merged PR 924438: Move changes to enable a Windows OS specific build.
These changes were made by Brian and have been active for a while in the kona_dialog branch. I've isolated the changes so can integrate only them into master to ease the operational load of keeping the other branch up to date and pay some of the cost of the merge in advance.

Related work items: #1630023