Merged PR 988661: There was a memory leak fix
There was memory leak inside the gstreamer extension as I was missing a buffer unmap. I added that. Now it fixes the leak issue.
On top of that I have reduced the sleep time to 2 ms as opposed to 10 ms to speed up the streaming.
Also get latest buffer size was not updating the correct size properly. I fixed it too.

Related work items: #1587739