Merged PR 1011913: Refactor for config and stream interops
Refactor for config and stream interops.
This could have been done for 1.3.0 release already, however was left out due to did not fit to schedule.
There is no need to have another layer inside Internal namespace and this simplifies the code when adding Conversation Transcriber with audio and video.

Related work items: #1741761