Merged PR 1015508: Add support for ARM64 unity builds.
The 1.4.0 release had undocumented support for ARM builds, but our Unity package did not yet have the ARM64 plug-in folder. Now that Unity 2019.1 is officially out, this change adds the ARM64 plug-in and documents beta support for ARM and ARM64 build types in our Unity quickstart readme.

Related work items: #1678269