Merged PR 1024845: [//build] Passed in encryped voice signature and its tag and its version per CELA request
In order to pass the CELA, Princeton is requested to encrypt the voice signatures. This requires passing in 3 pieces of information from client to service. We have decided it is easier for handing errors if Carbon takes a json string, instead of passing in a plain string with space separated 3 substrings.

Sara and I have tested with the client built from this  PR and her service locally, the service can return the expected user id. Her service has not been deployed yet. It will be pretty soon.

I am very sorry for this late PR and please let it. It is a must have by Princeton at Build.

I also added the wav header fix from AmitS into this PR. After my massive PR is in, it is safe now to add it.

Related work items: #1757073