Merged PR 1023815: [//build] Clang: put printf attribute on SpxTraceMessage(), fix warnings
Of the current compiler we're using, Clang seems to be most effective checking printf format strings and parameters (so, for the Android / mac / iOS targets).
This PR adds the printf attribute on SpxTraceMessage(), which we use internally in the core to log, and fixes all associated; these are latent bugs / potential crash issues that could surface in rare situations in customer side and/or limit our ability to debug customer issues (through log-to-file).

Related work items: #1760783