Merged PR 1031055: Take the elapsed time into account when deciding whether or not to throttle audio.
Take the elapsed time into account when deciding whether or not to throttle the audio pump.

When calculating if the thread processing audio needs to slow down, look at a previously calculated time for how soon the next audio packet should be queued and delay if needed.

Then calculate when the next packet should be queued.

Related work items: #1767480