Merged PR 1050437: Add one-parameter Java constructor for SpeechBotConnector
In C++ (and C#), we have a constructor for SpeechBotConnector that takes only one parameter--the config for the object. This is consistent with SpeechRecognizer and other objects. In Java, this one-parameter constructor is missing; that is, this doesn't work:

    SpeechBotConnector myJavaConnector = new SpeechBotConnector(myBotConfig); // constructor doesn't exist

This small change just adds the constructor chained to the existing one with AudioConfig.fromDefaultMicrophoneInput() as the audio config.

Related work items: #1694860