Merged PR 1065788: Fix ConversationPullStreamWithEnrollment() to tollerate results recieved out of time-sort order.
The service may return results out of order.
Long term, we should use a SortedList<long, string>() to sort the events based on the audio time the event is for.

For now, I am changing the test to only ensure the result was received, but not care what order the service returned the result in.

Related work items: #1782314