Merged PR 1070249: fix ItemPattern wild card list for the CheckSignatures build job.
The "itemPattern" seems to be in the wrong format for this task.
From Wei's log files, the task downloads the windows build artifact ok, but skips every file in the archive.  This causes the later steps to fail, when it can't even find the windows folder.

This changes this task to use the same multiple wild card format as the other uses of DownloadBuildArtifacts@) in our YML file.

Related work items: #1825992