Merged PR 1090044: Correct C# interop error from rename inconsistency
In the rename from SpeechBotConnector to DialogConnector, everything in speechapi_c_dialog_connector.cpp got "dialog_connector" prepended to it -- which was great, except for a couple of activity_received* methods, which received a new prefix instead of getting a replaced one.

    SPXAPI activity_received_event_release(SPXEVENTHANDLE h_event)

became

    SPXAPI dialog_connector_activity_received_event_release(SPXEVENTHANDLE h_event)

Which is good -- I like the consistency -- but we unfortunately didn't get the same in the interop layer for cs:

        [DllImport(Import.NativeDllName, CallingConvention = CallingConvention.StdCall)]
        internal static extern SPXHR activity_received_event_release(SPXEVENTHANDLE eventHandle);

...So you can guess what's happening.

Fix here is to just finish the consistency and add the dialog_connector (soon to be dialog_service_connector) to the interop.

Related work items: #1855559