Merged PR 1088840: Copy pma extension along with carbon libraries from NativeLibraryLoader.java
In the coming release of speech DDK, we are going to support Azure Kinect Microphone Array on Windows and Linux. On Windows, when we tried to test speech recognition end-to-end using Java project with Carbon and pma extension library, we noticed that Carbon was not using our audio processing no matter where we keep our library. To fix this, I have added some lines in `NativeLibraryLoader.java` to copy pma extension library to temp directory along with other Carbon libraries if it is present in the folder containing user's jar. If library cannot be found, Carbon will work as usual without printing any error messages or throwing any exceptions.

Related work items: #1853591