Merged PR 954181: Add debug support to dump recorded audio to file
A change to add ability to dump the audio sent to the service as well as the KWS for debug builds. The location of the audio dump files is controlled via string properties.

**USP audio:**
-----------------
**Property:** CARBON-INTERNAL-DumpAudioToDir
**Ex Value:** /sdcard/audiologs/
**Saved:** "<incrementingCount>_<impressionId>.wav"

The above will save 1 wav file (no RIFF header) for each service interaction. The file name contains the impression id.

**Kws audio:**
-----------------
For KWS audio we support dumping all audio processed by KWS to a circular list of files. The following will write to 5 files each with 20 seconds worth of audio. Older files will be overwritten.

**Property:** CARBON-INTERNAL-KwsDumpAudioToDir
**Ex Value:** /sdcard/kws-continuous/

**Property:**  (optional) CARBON-INTERNAL-KwsDumpAudioCircularFileLengthMs
**Ex Value:** "20000"

**Property:**  (optional)  CARBON-INTERNAL-KwsDumpAudioCircularFileCount
**Ex Value:** "5"

**Saved:** "continuous_kws_<0..4>.wav"

Related work items: #1833476