PostAsJson leads to HTTP 400 from Speech API
There seemed to be an issue with using HttpClient.PostAsJsonAsync with the json formatting or mediatype encoding. Speech API was replying 400, Cannot find the specified object
Manually serializing the transcription to json and calling HttpClient.PostAsync solved the issue in my case.

Also, there seems to me there is no point in setting the res variable and then calling PostAsJson. This is why I am suggesting a change anyway