Merged PR 1108855: Fix for Bug 1717575: Linux C++ Compressed Audio Sample: hang for missing
Fix for Bug 1717575: Linux C++ Compressed Audio Sample: hang for missing subscription info.

Gstreamer audio decoding happens inside gstreamer code. Due to some external factor when there is some error (like subscription key missing or some auth error) the session closes the pump and the pump thread gets stopped but we were not notifying the gstreamer adapter that the pump is closed. So at the time of stopping the pump we need to close the adapter and which will make sure that the reader is closed properly and all the gstreamer resources will be released properly.

Related work items: #1717575