Merged PR 1113519: update build scripts to support new Unity build agent.
Our old UnityBuild agent is old and very low on disk space.
Nick built a new Agent and also updated the versions of Unity and Visual Studio installed, so that we can build ARM64 unity packages when we need to.

These are the needed changes to work with the new on-prem build agent for unity.

Related work items: #1777956