Merged PR 1120290: Added test coverage for speech context json logic
As part of language ID effort, we need to update speech context json logic.  This includes 2 efforts, 1st is to replace raw string json implementation with json library, so we deal with json objects instead of raw strings.  2nd is to extend the speech context json to support language ID feature.  For both efforts, we need to make sure test coverage is in place so we have confidence not breaking the json format.  This change is to add this test coverage.  I did a small refinement to put all context json generation logic to GetSpeechContextJson method.  This is to facilitate test implementation, also to avoid GetSpeechContextJson having longer and longer parameter list.

Related work items: #1701318, #1894493