Merged PR 1112941: commit b300b32a: bug fix for Linux C++ Compressed Audio Sample: bad error if
commit b300b32a: bug fix for Linux C++ Compressed Audio Sample: bad error if wrong audio format is specified.

There was a error assignment missing in case of error and which was causing this error.

With the current fix most of the errors should be fixed.

But one scenario in older gstreamer version is still there with oggdemux. When we give a wrong stream to oggdemux, they do not raise any error. But that was fixed with later version of Gstreamer. So we still have issue with Ubuntu 16.04 only for ogg demux.

Related work items: #1717577