Merged PR 1116879: adding cgmanifest entry for all public samples.
We had a minor usage of gradle before, and I am moving it to a new cgmanifest file for the samples folder.
Nick added usage for TINY JSON, and this ships with the unity sample as a code snippet.
This change adds our new usage of this existing OSS component.

Related work items: #1894218