Merged PR 1141871: Fix occasional java vm crash on osx
We saw java vm crashed due to sigtrap(6) at shutdown time on osx. This is probably caused by the fact that event callback might still be triggered during shutdown time. This PR is to prevent populating any callbacks to audio session layer if it is in closing phase.
I have run the test after for >100 times in CI build, and have not seen any crash. Below are the latest tests runs:
https://msasg.visualstudio.com/Skyman/_build/results?buildId=6155887
https://msasg.visualstudio.com/Skyman/_build/results?buildId=6155885
https://msasg.visualstudio.com/Skyman/_build/results?buildId=6155875
https://msasg.visualstudio.com/Skyman/_build/results?buildId=6155824
https://msasg.visualstudio.com/Skyman/_build/results?buildId=6155819
https://msasg.visualstudio.com/Skyman/_build/results?buildId=6155813
https://msasg.visualstudio.com/Skyman/_build/results?buildId=6155810
https://msasg.visualstudio.com/Skyman/_build/results?buildId=6155807

Related work items: #1887564