Merged PR 1096792:  CSpxThreadService::ExecuteSync improvement
The promise of executed should signal after thread service is completely done with the task. That is after destructor is calledm, aka task is out of scope, and after finish running the task.

This is the root cause for failure if conversation_id in conversation_online_pull_stream in cxx_api_tests

I have turned on the conversation_online_pull_stream and have 6 build jobs, all successes in above test case.

https://msasg.visualstudio.com/Skyman/_build/results?buildId=5881327
https://msasg.visualstudio.com/Skyman/_build/results?buildId=5881326
https://msasg.visualstudio.com/Skyman/_build/results?buildId=5866411
https://msasg.visualstudio.com/Skyman/_build/results?buildId=5861453 -->red due to other tests failed
https://msasg.visualstudio.com/Skyman/_build/results?buildId=5859342
https://msasg.visualstudio.com/Skyman/_build/results?buildId=5859338 -->red due to other tests failed

Related work items: #1859511