Merged PR 1148025: Alaw, Mulaw and FLAC
Adding Alaw,mulaw,flac support.

Abandoned the previous PR as it was really difficult to remove the merge conflict. This PR has support for Alaw,MuLAw and flac for Android and iOS Also.

I was planning to add amrnb and amrwb support also. I have put the code for that too. But somehow I was not able to create test stream to test the amrnb and amrwb codec. So I decided not to release it now. We can look at this when a demand comes for it. I have commented the code for it. We need to uncomment code related to amr and it should work (but currently not tested)

Related work items: #1897548, #1898352, #1898353, #1902430, #1902434