Merged PR 1169728: Long running tests are timeing out, but that also mean we get zero logs to
Long running tests are timeing out in nightly builds, but that also mean we get zero logs to diagnose the problem.  I'm bumping the timeout so we can investigate what in LongRunning tests is hung or just failing and causing retry to got past the max build timeout allowed.

Related work items: #1902581