FIX: Speech API is replying "400, Cannot find the specified object" (#296)
* null does not get serialized as an empty array

sending null won't let the json serializer emit an empty array for the modelIds. null will be used instead and this is agains the API specs

* PostAsJson leads to HTTP 400 from Speech API

There seemed to be an issue with using HttpClient.PostAsJsonAsync with the json formatting or mediatype encoding. Speech API was replying 400, Cannot find the specified object
Manually serializing the transcription to json and calling HttpClient.PostAsync solved the issue in my case.

Also, there seems to me there is no point in setting the res variable and then calling PostAsJson. This is why I am suggesting a change anyway