Merged PR 1156293: Linux-ARM64 build support
Added Linux-arm64 to supported Carbon and KWS build targets.
Updated CI configuration to build and test Carbon with KWS on Linux-arm64.
Updated the KWS library build id to a newer version that provides corresponding "Linux_AARCH64" libraries.
Added Linux-arm64 binaries to nuget, LinuxDrop tar.gz, and Java jar packages.

With the initial changes, "Dialog Service Connector KWS basics" in cxx_api_tests failed consistently on Linux-arm64 in CI, with SPXERR_SETFORMAT_UNEXPECTED_STATE_TRANSITION, but passed when run locally in a Linux-arm64 container. The reason was that in CI the audio pump thread reached the end of input stream before the keyword was detected, and called SetFormat with nullptr. But this would only be handled in a background task that was created but didn't get to run until after the keyword had been detected and session state changed to HotSwapPaused. This format and state combination was not among the supported cases in SetFormat and triggered the error.

To solve the issue, I updated SetFormat to accept the described condition but only set a flag, and the processing of buffered audio leftovers is then triggered based on the flag after the hot swap is over. During investigation it was also necessary to add some more traces to help debugging and make the logs more readable (at least to me), most of them are included in the change set.

Related work items: #1840913