Merged PR 1181568: fix bug causing segfault test failures on LinuxARM32
Nightly builds have been failing in the last 14 days, due to a test failure on LinuxARM32 (which only nightly). It was caused by a bug added to the code recently. This PR fixes the bug.

I manually run LinuxARM32 test locally as well as in VSO. No failures so far.

Related work items: #1951819