Merged PR 1247737: Remove typed Activity classes from the DialogServiceConnector APIs
The original implementations in the Dialog namespace include a typed representation of Bot Framework Activities in the C++ layer to simplify development; the BF SDKs are not available in Cpp. This creates inconsistency between the projections and introduces unnecessary extra complexity. Here we remove all the layers and just treat Activities as text.

Related work items: #1969970