Merged PR 1269688: Add nuget target definitions for cpp UWP support
Reviving from a commit that Eric shared *last year* -- but we extra especially need it now with the Modern Dictation work going on.

**Problem**: A UWP C++ application referencing a C++ library that in turn uses Carbon (via nuget) *does not* get the needed DLLs copied into the final appx folder; they stay uselessly stranded in the secondary project binaries folder. This makes it impossible (or at least very manual and hacky) to get an automated CI process with this project configuration working. The project configuration is exactly what Modern Dictation is doing, so this is a problem.

Eric drafted up this .targets fix as a tentative solution last year but we never got to test/validate it. Now let's see if we can't get it in for the next release.

Related work items: #2033491