Merged PR 1237744: When releasing a recognizer call recognizer_stop_continuous_recognition first
When releasing a recognizer call recognizer_stop_continuous_recognition first to attempt to put the recognizer in a known state.

A running recognizer risks being in a worker block when the object is released that will hold a handle to the recognizer. This will result in an unhandled exception as the thread pool cannot be closed internally.

Related work items: #1842187