Merged PR 1290278: Removed mode from LID speech.context per discussion with Shruti
The mode is optional and if we don't set it,  service uses whatever the endpoint's mode is. This is the behavior we want.  In future, if we have one service endpoint that can support multiple modes, we can use this json property to differentiate which mode the user scenario is.  Also removed the mode from phraseOutput and phraseDectation section, as it is duplicate information with 'mode' json property.

Related work items: #2002919