Merged PR 1290533: Fix the StartPump / StopPump race condition in CSpxAudioStreamSession + detailed tracing
This is the result of an investigation of irecoverable state in Android
apps when using the USB mic array is used. There are multiple root
causes leading to this, this change fixes one of them. In cases of quick
succession from starting the recogntion (using the dialogServiceConnector)
to closing the object while the recognition is still starting the pump is
not closed and is leaked.

1. Modified Term to wait for the transition into ProcessingAudio state and if
not close the Pump even if WaitForPumpSetFormatStart.
2. Detailed tracing into CSpxAudioStreamSession and CSpxAudioPump so we can
trace objects lifetime closely.

Related work items: #2057910