Merged PR 1304117: Speed up VS builds
This turns on multi-processor compilation for all of the generated .vcxproj files in Visual Studio and/or msbuild to speed up builds. Without this flag, the build process seems to compile a single .cpp at a time which is quite slow and doesn't take full advantage of your CPU (on my machine <10% was used). With this flag, cl.exe will batch up files within a project that use the same command line arguments and compile them in parallel. See here for a more in depth explanation:
https://randomascii.wordpress.com/2014/03/22/make-vc-compiles-fast-through-parallel-compilation/

On my machine, rebuilding the carbon.sln from the command line decreased from 6min 45s to 1min 55sec by turning this on. Turning on project level parallelism as well (/m msbuild switch) further reduced rebuild time to 1min 18sec. Of course your mileage may vary.

To take full advantage of the faster build:
- Make sure you re-run the "cmake -G "Visual Studio 15 Win64" .." in your build folder to update all the project files
- Visual studio seems to parallelise at the project level on its own so nothing extra is needed here
- If you are building from the command line, make sure you include the /m parameter. E.g. msbuild /m carbon.sln

Related work items: #1880149