Merged PR 1311939: Add ability to overflow on write into the ring buffer
The ring buffer has two pointers moving, the read pointer and write pointer. The read pointer works fine when only one reader is present as this pointer is tracked by the buffer itself. With multiple readers each one keeps the read position by itself and the buffer tracked read pointer is not moving affecting the overflow buffer decisions later on. The code allows for the read pointer to be carried along on a given setting such that writer will not be blocked but override old data. The expectation is that readers will move along lively otherwise they will start getting errors if they lag behind.

* Control the overflow through AudioSourceBufferAllowOverflow property
* Added unit tests to validate the new behavior in the ring buffer

Related work items: #2092172