Merged PR 1349249: Added lidEnabled=true if lid config is found
Added lidEnabled=true if lid config is found.  this is a requirement for 3P, as 3P has some regions not enabled NewMan graph yet and LID feature is only on NewMan graph.  So 3P service need this query parameter to know if it needs to initiate NewMan graph.

Related work items: #2152775