Merged PR 1367593: More breaking changes to expected service results.
Some minor changes to SR and Translation results push the Long Running tests past the threshold of their fuzzy matching checks. This PR updates the expects results so the tests are back in sync with the actual service results.
This is blocking nightly builds, as retrying the long running tests runs over the max build time.

Related work items: #2170907