Merged PR 1370923: Added lid only support for AMS team, 1st party only
AMS team has scenario that only need language detection, without speech recognition.  This change added a hard coded string property to enable such a mode. When this property is set to true, Carbon send speech.context telling service the client doesn't want speech recognition.

This feature is 1st party only for now, so we are not adding a new API for it.

Related work items: #2175763